began development of administrator mode
