+++ src/app/main/include/main.h
@@ -25,17 +25,20 @@
 #include "display_config.h"
 #include "display_frames.h"
 #include "pir.h"
+#include "kiosk_tasks.h"
 
 // MACROS
-#define BLINK_GPIO 8
-#define MAIN_DEBUG 1
-#define PROXIMITY_DETECTED_BIT BIT0
-#define ID_ENTERED_SUCCESS_BIT BIT1
-#define ID_AUTHENTICATED_BIT BIT2
-#define IDLE_BIT BIT4
-#define ENTERING_ID_BIT BIT5
-#define TAG "MAIN"
-#define ID_LEN 7
+// #define BLINK_GPIO 8
+// #define MAIN_DEBUG 1
+// #define PROXIMITY_DETECTED_BIT BIT0
+// #define ID_ENTERED_SUCCESS_BIT BIT1
+// #define ID_AUTHENTICATED_BIT BIT2
+// #define ADMIN_MODE_BIT BIT3
+// #define IDLE_BIT BIT4
+// #define ENTERING_ID_BIT BIT5
+// #define NEW_ID_ENTERED_SUCCESS_BIT BIT6
+// #define TAG "MAIN"
+// #define ID_LEN 7
 
 // STRUCTS
 typedef enum
@@ -45,20 +48,12 @@ typedef enum
     STATE_ERROR,
 } kiosk_state_t;
 
-typedef enum
-{
-    STATE_IDLE,
-    STATE_USER_DETECTED,
-    STATE_VALIDATING,
-    STATE_DISPLAY_RESULT
-} state_t;
-
 // GLOBAL VARIABLES
 // TASK HANDLES
 static TaskHandle_t blink_led_task_handle = NULL;
 static TaskHandle_t wifi_init_task_handle = NULL;
 static TaskHandle_t ota_update_task_handle = NULL;
-static TaskHandle_t keypad_task_handle = NULL;
+TaskHandle_t keypad_task_handle = NULL;
 static TaskHandle_t lvgl_port_task_handle = NULL;
 
 // not static because it is being used in wifi_init.c as extern variable
@@ -67,30 +62,21 @@ EventGroupHandle_t event_group;
 
 // STATE VARIABLES
 static kiosk_state_t current_kiosk_state = STATE_WIFI_INIT;
-static state_t current_state = STATE_IDLE, prev_state = STATE_IDLE;
+state_t current_state = STATE_IDLE, prev_state = STATE_IDLE;
 
 // EXTERN VARIABLES
-pn532_t nfc;                   // Defined in ntag_reader.c
-keypad_buffer_t keypad_buffer; // Defined in keypad_driver.c
-_lock_t lvgl_api_lock;         // Defined in display_config.c
-lv_display_t *display;         // Defined in display_config.c
+pn532_t nfc;                   // Defined in ntag_reader.h
+keypad_buffer_t keypad_buffer; // Defined in keypad_driver.h
+_lock_t lvgl_api_lock;         // Defined in display_config.h
+lv_display_t *display;         // Defined in display_config.h
+lv_obj_t *disp_obj;            // Defined in kiosk_states.h
 
-static char nfcUserID[MAX_ID_LEN];    // Used to store the ID read from NFC
-static char keypadUserID[MAX_ID_LEN]; // Used to store the ID entered from the keypad
-bool idIsValid = true;                // Flag set by database query results
-bool keypadEnteredFlag = false;
-bool nfcReadFlag = false;
 static led_strip_handle_t led_strip;
 static uint8_t s_led_state = 0;
-static lv_obj_t *disp_obj;
 
 // FUNCTION PROTOTYPES
 static void blink_led(void);
 static void configure_led(void);
-void proximity_task(void *param);
-void nfc_scan_id_task(void *param);
-void nfc_enter_id_task(void *param);
-void validation_task(void *param);
-void display_task(void *param);
+
 
 #endif
